feat: add createUser and getUserByIf functions
