+++ src/userOperatons/getUsers.js
@@ -1,3 +1,5 @@
-export const handleGetUsers = (res) => {
+import users from "../utils/users.json";
+
+export const getUsers = (res) => {
   sendJsonResponse(res, 200, users);
 };
